pyproject.toml
[project]
name = "classification-of-reviews"
version = "0.1.0"
description = "classification of reviews from online store"
authors = [
    {name = "Vladislav Ulandaev", email = "ulandaev.vn@phystech.edu"}
]
readme = "README.md"
requires-python = "~3.12"

[tool.poetry.dependencies]
pandas = ">=2.3.2,<3.0.0"
torch = ">=2.8.0,<3.0.0"
numpy = "^2.3.3"
matplotlib = "^3.10.6"
scikit-learn = "^1.7.2"
transformers = "^4.56.2"
datasets = "^4.1.1"
mlflow = "^3.4.0"
peft = "^0.17.1"


[tool.poetry.group.dev.dependencies]
fire = ">=0.7.1,<0.8.0"
ipykernel = "^5.5.0"
dvc = "^1.11.16"
pre-commit = "^2.10.1"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

